refactor: task adapter
