+++ envs/mini_swe_env/__init__.py
@@ -1,51 +1,32 @@
-"""Mini SWE environment — SWE-bench task adapter + environment.
+"""Mini SWE environment — task models + environment.
 
 Exports:
-    Task adapter (Phase 2):
-        SWETask, validate_swe_task, adapt_swebench_lite_row, ...
+    Task models:
+        SWETask, validate_swe_task, coerce_swe_task, SWETaskValidationError
 
-    Environment + client (Phase 3):
+    Environment + client:
         SWEEnvironment, MiniSWEEnv, SWERolloutResult, ...
 """
-
-from .task_loader_swebench_lite import (
-    adapt_swebench_lite_row,
-    adapt_swebench_lite_rows,
-    AdaptationSkip,
-    coerce_swe_task,
-    deterministic_train_eval_split,
-    load_task_file,
-    read_jsonl_rows,
-    SWEBenchLiteAdapterError,
-    SWETask,
-    SWETaskValidationError,
-    validate_swe_task,
-    write_tasks_jsonl,
-)
 
 from .models import (
     SWECommandResult,
     SWERolloutResult,
     SWEState,
+    SWETask,
+    SWETaskValidationError,
+    coerce_swe_task,
+    validate_swe_task,
 )
 
 from .client import MiniSWEEnv
 
 __all__ = [
-    # Task adapter (Phase 2)
-    "AdaptationSkip",
-    "SWEBenchLiteAdapterError",
+    # Task models
     "SWETask",
     "SWETaskValidationError",
-    "adapt_swebench_lite_row",
-    "adapt_swebench_lite_rows",
     "coerce_swe_task",
-    "deterministic_train_eval_split",
-    "load_task_file",
-    "read_jsonl_rows",
     "validate_swe_task",
-    "write_tasks_jsonl",
-    # Models (Phase 3)
+    # Models
     "MiniSWEEnv",
     "SWECommandResult",
     "SWERolloutResult",
